% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_gene_percentage}
\alias{get_gene_percentage}
\title{Internal helper function 1}
\usage{
get_gene_percentage(expr_matrix, pattern, total_counts)
}
\description{
These functions support QCseq’s internal calculations.
They are not exported to users.
}
\keyword{internal}
