% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCcompute.R
\name{QCcompute}
\alias{QCcompute}
\title{Computing independent quality control metrics}
\usage{
QCcompute(expr_matrix)
}
\arguments{
\item{expr_matrix}{A Seurat object, RNA assay, gene expression matrix
where columns are cells and rows are genes.}
}
\value{
Returns a dataframe where rows are Cells, and columns are:
number of genes (n_genes), number of transcripts (n_transcripts),
percentage of mitochondrial genes (percent_mit), and percentage of ribosomal
genes (percent_ribo).
}
\description{
Computes the following quality control metrics from a scRNA-seq gene
expression matrix: 1) number of detected genes per cell, 2) total transcripts
per cell, 3) percentage of mitochondrial genes, and 4) percentage of
ribosomal genes.
}
\examples{
# Using sample_scRNAseq dataset available with the package
data("sample_scRNAseq") # Access sample data
qc_results <- QCcompute(sample_scRNAseq) # Compute QC metrics for each sample
head(qc_results) # Display results head

}
\references{
Hao, Y., Stuart, T. A., Kowalski, M. H., Choudhary, S., Hoffman, P., Hartman,
A., Srivastava, A., Molla, G., Shaista Madad, Fernandez-Granda, C., & Rahul
atija. (2023). Dictionary learning for integrative, multimodal and scalable
single-cell analysis. Nature Biotechnology, 42, 293–304.
https://doi.org/10.1038/s41587-023-01767-y

Lu, J., Sheng, Y., Qian, W., Pan, M., Zhao, X., & Ge, Q. (2023).
scRNA‐seq data analysis method to improve analysis performance.
Iet Nanobiotechnology, 17(3), 246–256. https://doi.org/10.1049/nbt2.12115.

OpenAI. (2025). ChatGPT. ChatGPT 5; OpenAI. https://chatgpt.com/.

R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.
}
