% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QChistogram.R
\name{QChistogram}
\alias{QChistogram}
\title{Returns a histogram of quality control scores across all samples.}
\usage{
QChistogram(expr_matrix, bar_colour = "#4C72B0")
}
\arguments{
\item{expr_matrix}{A Seurat object, RNA assay, gene expression matrix
where columns are cells and rows are genes.}

\item{bar_colour}{Hex code for colour to allow users to customize the colour
of the bars in the histogram, default is #4C72B0 (light blue)}
}
\value{
Returns ggplot histogram plot.
}
\description{
Returns a publication-level histogram where x-axis is quality control score
(from QCscore function) and y-axis is frequency. Users can customize colour.
}
\examples{
# Using sample_scRNAseq dataset available with the package
data("sample_scRNAseq") # Access sample data
QChistogram(sample_scRNAseq) # Output histogram for QC scores

}
\references{
R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.
}
