% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCpca.R
\name{QCpca}
\alias{QCpca}
\title{PCA plot labeled by QC score}
\usage{
QCpca(expr_matrix)
}
\arguments{
\item{expr_matrix}{A Seurat object, RNA assay, gene expression matrix
where columns are cells and rows are genes.}
}
\value{
Returns ggplot PCA scatter plot labeled by QC rank score.
}
\description{
The function performs a principal component analysis (PCA) on key scRNAseq
QC metrics including nGenes, nTranscripts, percent_mit, and percent_ribo.
PCA is then applied to numeric QC features to reduce the data to 2 main
dimensions (PC1 and PC2), which capture the greatest sources of variation
among the cells.
}
\examples{
# Using sample_scRNAseq dataset available with the package
data("sample_scRNAseq") # Access sample data
QCpca(sample_scRNAseq) # Output PCA plot based on QC scores

}
\references{
R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.
}
