% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCscore.R
\name{QCscore}
\alias{QCscore}
\title{Computing unified quality control score}
\usage{
QCscore(expr_matrix)
}
\arguments{
\item{expr_matrix}{A Seurat object, RNA assay, gene expression matrix
where columns are cells and rows are genes.}
}
\value{
Returns a dataframe where rows are Cells, and columns are:
number of genes (nGenes), number of transcripts (nTranscripts),
percentage of mitochondrial genes (percent_mit), percentage of ribosomal
genes (percent_ribo), and unified QC score (qc_score_rank).
}
\description{
Computes a unified quality control score from 3 independent QC metrics
(nGenes, nTranscripts, and percent_mit) using median absolute deviation (MAD)
standardization and rank-based scores.
}
\examples{
# Using sample_scRNAseq dataset available with the package
data("sample_scRNAseq") # Access sample data
qc_results_with_score <- QCscore(sample_scRNAseq) # QC metrics + Score
head(qc_results_with_score) # Display results head

}
\references{
Bacher, R., Chu, L.-F., Argus, C., Bolin, J. M., Knight, P., Thomson, J.,
Stewart, R., & Kendziorski, C. (2021). Enhancing biological signals and
detection rates in single-cell RNA-seq experiments with cDNA library
equalization. Nucleic Acids Research, 50(2), e12–e12.
https://doi.org/10.1093/nar/gkab1071.

OpenAI. (2025). ChatGPT. ChatGPT 5; OpenAI. https://chatgpt.com/.

R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.
}
