% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCsummary.R
\name{QCsummary}
\alias{QCsummary}
\title{Return a clear summary report for the gene expression data.}
\usage{
QCsummary(expr_matrix, threshold = 0.5)
}
\arguments{
\item{expr_matrix}{A Seurat object, RNA assay, gene expression matrix
where columns are cells and rows are genes.}

\item{threshold}{Tpye numeric, sets the QC score threshold for what is an
acceptable score. Deafult = 0.5.}
}
\value{
Returns text (type character). It is a clear quality control summary
report specific to the quality control metrics extracted form the scRNAseq
gene expression matrix.
}
\description{
Returns a clear summary report for the gene expression data using the key
quality control metrics (nGenes, nTranscripts, percent_mit, and percent_ribo)
and the signature QC score that combines the metrics.
}
\examples{
# Using sample_scRNAseq dataset available with the package
data("sample_scRNAseq") # Access sample data
qc_summary <- QCsummary(sample_scRNAseq) # Output QC summary paragraph
qc_summary # Display the output paragraph

}
\references{
R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.
}
