% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{robust_scale}
\alias{robust_scale}
\title{Internal helper function 2}
\usage{
robust_scale(x)
}
\description{
These functions support QCseq’s internal calculations.
They are not exported to users.
}
\keyword{internal}
