% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQCseq.R
\name{runQCseq}
\alias{runQCseq}
\title{Launch Shiny App for QCseq}
\usage{
runQCseq()
}
\value{
launches the Shiny app
}
\description{
This functions launches the Shiny app for QCseq The app is intended for
those with less coding experience and performs the entire QCseq workflow.
}
\examples{
\dontrun{
QCseq::runQCseq()
}

}
\references{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Aden-Buie G, Xie Y,
Allen J, McPherson J, Dipert A, Borges B (2025). shiny: Web Application
Framework for R. R package version 1.12.0, https://shiny.posit.co/.

R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.

Shiny - File Upload. (2014, July 29). Shiny.
https://shiny.posit.co/r/gallery/widgets/file-upload/

Shiny - Including HTML, text, and Markdown files. (2014, July 29). Shiny.
https://shiny.posit.co/r/gallery/application-layout/including-html-text-and-markdown-files/

Shiny - Tabsets. (2014, July 30). Shiny.
https://shiny.posit.co/r/gallery/application-layout/tabsets/
}
