% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_scRNAseq}
\alias{sample_scRNAseq}
\title{Single-cell RNA sequencing (scRNAseq) gene counts}
\format{
A Seurat object with 13,714 features (genes) across 2,700 samples
(cells) within 1 assay
}
\source{
10x Genomics
}
\usage{
sample_scRNAseq
}
\description{
A dataset of Peripheral Blood Mononuclear Cells (PBMC) freely available from
10X Genomics. There are 2,700 single cells that were sequenced on the
Illumina NextSeq 500
}
\examples{
\dontrun{
sample_scRNAseq
}
}
\keyword{datasets}
