% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCsummary.R
\name{QCsummary}
\alias{QCsummary}
\title{Return a clear summary report for the gene expression data.}
\usage{
QCsummary(expr_matrix, lower_threshold = 0.4, upper_threshold = 0.75)
}
\arguments{
\item{expr_matrix}{A Seurat object, RNA assay, gene expression matrix
where columns are cells and rows are genes.}

\item{lower_threshold}{Type numeric, default = 0.4. If the mean of all QC
scores is below this value, the dataset quality is poor.}

\item{upper_threshold}{Type numeric, default = 0.75. If the mean of all QC
scores is above this value, the dataset quality is good. If the mean of all
QC scores is between 0.4 and 0.75, the dataset quality is moderate.}
}
\value{
Returns text (type character). It is a clear quality control summary
report specific to the quality control metrics extracted form the scRNAseq
gene expression matrix.
}
\description{
Returns a clear summary report for the gene expression data using the key
quality control metrics (n_genes, n_transcripts, percent_mit, and percent_ribo)
and the signature QC score that combines the metrics.
}
\examples{
# Using sample_scRNAseq dataset available with the package
data("sample_scRNAseq") # Access sample data
qc_summary <- QCsummary(sample_scRNAseq) # Output QC summary paragraph
qc_summary # Display the output paragraph

}
\references{
R Core Team (2025). \emph{R: A Language and Environment for Statistical
Computing}. R Foundation for Statistical Computing, Vienna, Austria.
https://www.R-project.org/.
}
